Review comments: rework the test
