+++ integrations/openapi-ui/src/it/projects/test1/src/main/java/module-info.java
@@ -13,7 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 /**
- * Integration test for OpenAPI UI options set programmatically.
+ * Test module.
  */
-package io.helidon.tests.integration.openapiui.module;
+module it {
+
+    requires io.helidon.integrations.openapi.ui;
+}
