--- tests/integration/gh-10889-openapi-ui-module/helidon-tests-integration-openapi-ui-module/src/main/java/io/helidon/tests/integration/openapiui/module/Main.java
@@ -1,86 +1,0 @@
-/*
- * Copyright (c) 2025 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.helidon.tests.integration.openapiui.module;
-
-
-import io.helidon.integrations.openapi.ui.OpenApiUi;
-import io.helidon.logging.common.LogConfig;
-import io.helidon.config.Config;
-import io.helidon.openapi.OpenApiFeature;
-import io.helidon.webserver.WebServer;
-import io.helidon.webserver.http.HttpRouting;
-
-
-
-
-/**
- * The application main class.
- */
-public class Main {
-
-
-    /**
-     * Cannot be instantiated.
-     */
-    private Main() {
-    }
-
-
-    /**
-     * Application main entry point.
-     * @param args command line arguments.
-     */
-    public static void main(String[] args) {
-        
-        // load logging configuration
-        LogConfig.configureRuntime();
-
-        // initialize global config from default configuration
-        Config config = Config.create();
-        Config.global(config);
-
-        var openApiConfig = config.get("openapi");
-        var openApiUiConfig = openApiConfig.get("ui");
-        var ui = OpenApiUi.builder()
-                .config(openApiUiConfig)
-                .webContext("/my-ui")
-                .build();
-
-        WebServer server = WebServer.builder()
-                .config(config.get("server"))
-                .addFeature(OpenApiFeature.builder()
-                                    .addService(ui)
-                                    .build())
-                .config(config.get("openapi"))
-                .routing(Main::routing)
-                .build()
-                .start();
-
-
-        System.out.println("WEB server is up! http://localhost:" + server.port() + "/simple-greet");
-
-    }
-
-
-    /**
-     * Updates HTTP Routing.
-     */
-    static void routing(HttpRouting.Builder routing) {
-        routing
-               .register("/greet", new GreetService())
-               .get("/simple-greet", (req, res) -> res.send("Hello World!")); 
-    }
-}
