--- tests/integration/gh-10889-openapi-ui-module/helidon-tests-integration-openapi-ui-module/src/main/resources/logging.properties
@@ -1,20 +1,0 @@
-#
-# Copyright (c) 2025 Oracle and/or its affiliates.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-handlers=io.helidon.logging.jul.HelidonConsoleHandler
-java.util.logging.SimpleFormatter.format=%1$tY.%1$tm.%1$td %1$tH:%1$tM:%1$tS.%1$tL %4$s %5$s%6$s%n
-# Global logging level. Can be overridden by specific loggers
-.level=INFO
-io.helidon.level=INFO
