--- tests/integration/gh-10889-openapi-ui-module/helidon-tests-integration-openapi-ui-module/src/test/java/io/helidon/tests/integration/openapiui/module/AbstractMainTest.java
@@ -1,69 +1,0 @@
-/*
- * Copyright (c) 2025 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.helidon.tests.integration.openapiui.module;
-
-
-import io.helidon.common.media.type.MediaTypes;
-import io.helidon.http.Status;
-import io.helidon.webclient.api.ClientResponseTyped;
-import io.helidon.webclient.http1.Http1Client;
-import io.helidon.webclient.http1.Http1ClientResponse;
-import io.helidon.webserver.http.HttpRouting;
-import io.helidon.webserver.testing.junit5.SetUpRoute;
-
-import org.junit.jupiter.api.Test;
-import jakarta.json.JsonObject;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.CoreMatchers.containsString;
-
-abstract class AbstractMainTest {
-    private final Http1Client client;
-
-    protected AbstractMainTest(Http1Client client) {
-        this.client = client;
-    }
-
-    @SetUpRoute
-    static void routing(HttpRouting.Builder builder) {
-        Main.routing(builder);
-    }
-
-    @Test
-    void testGreeting() {
-        ClientResponseTyped<JsonObject> response = client.get("/greet").request(JsonObject.class);
-        assertThat(response.status(), is(Status.OK_200));
-        assertThat(response.entity().getString("message"), is("Hello World!"));
-    }
-
-    
-    @Test
-    void testMetricsObserver() {
-        try (Http1ClientResponse response = client.get("/observe/metrics").request()) {
-            assertThat(response.status(), is(Status.OK_200));
-        }
-    }
-
-    
-    @Test
-    void testSimpleGreet() {
-        ClientResponseTyped<String> response = client.get("/simple-greet").request(String.class);
-        assertThat(response.status(), is(Status.OK_200));
-        assertThat(response.entity(), is("Hello World!"));
-    }
-
-}
